pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prgx-ag-nexus"
version = "0.1.0"
description = "PRGX-AG Nexus backend for Porisjem Protocol orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6,<3",
  "pydantic-settings>=2.2,<3",
  "PyYAML>=6.0,<7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"
